ACC-3028 move DataEncryptionKeyTableCreator to autoconfigure
This removes the dependency from contentgrid-appserver-contentstore-impl-encryption to contentgrid-appserver-query-engine-impl-jooq
