+++ contentgrid-appserver-autoconfigure/src/test/java/com/contentgrid/appserver/autoconfigure/contentstore/EncryptedContentStoreAutoConfigurationTest.java
@@ -3,7 +3,6 @@ package com.contentgrid.appserver.autoconfigure.contentstore;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.contentgrid.appserver.autoconfigure.query.engine.JOOQQueryEngineAutoConfiguration;
-import com.contentgrid.appserver.contentstore.impl.encryption.keys.DataEncryptionKeyTableCreator;
 import com.contentgrid.appserver.contentstore.impl.fs.FilesystemContentStore;
 
 import com.contentgrid.appserver.query.engine.jooq.JOOQQueryEngine;
