+++ contentgrid-appserver-contentstore-impl-encryption/src/main/java/com/contentgrid/appserver/contentstore/impl/encryption/keys/TableStorageDataEncryptionKeyAccessor.java
@@ -26,7 +26,7 @@ public class TableStorageDataEncryptionKeyAccessor implements DataEncryptionKeyA
     private static final Field<String> ALGORITHM = field(name(TABLE_NAME, "algorithm"), SQLDataType.VARCHAR);
     private static final Field<byte[]> INITIALIZATION_VECTOR = field(name(TABLE_NAME, "iv"), SQLDataType.BLOB);
 
-    static void setupTables(DSLContext dslContext) {
+    public static void setupTables(DSLContext dslContext) {
         dslContext.createTableIfNotExists(TABLE_NAME)
                 .column(CONTENT_ID)
                 .column(KEK_LABEL)
@@ -37,7 +37,7 @@ public class TableStorageDataEncryptionKeyAccessor implements DataEncryptionKeyA
                 .execute();
     }
 
-    static void dropTables(DSLContext dslContext) {
+    public static void dropTables(DSLContext dslContext) {
         dslContext.dropTableIfExists(TABLE_NAME).execute();
     }
